data-engineering-projects/room_occupancy_detection_g5-main/luxmeter_extractor/requirements.txt
APScheduler==3.10.0
kafka-python==2.0.2
requests==2.28.2
